requirements.txt
openai==0.27.8
pandas==1.5.2
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.1